+++ server/tests/integration.rs
@@ -32,3 +32,15 @@ async fn user_crud() {
   send(&c, ChangePasswordPayload::default(), "PUT", "users/change-password", 200, "d").await;
   send(&c, "", "GET", "users/alice", 200, "e").await;
 }
+
+
+#[tokio::test]
+#[serial]
+async fn items_crud() {
+  let mut _child_guard = cargo_shuttle_run().await;
+  let c = Client::builder().cookie_store(true).build().unwrap();
+  send(&c, UserPayload::default(), "POST", "users", 200, "1").await;
+  send(&c, CredentialsPayload::default(), "POST", "users/login", 200, "2").await;
+  // send(&c, )
+
+}
